Add subgraph support to manifests (#65)
This includes manifest changes to edge definitions. Now there are `local_edges` and `resolved_edges`, which both have same same format.

Due to the way nodes are specified in the yaml, the algorithm for assigning edges is a bit complicated.

Because graph nodes have edges that connect to both child and parent nodes to arbitrary depth, and their interface is dependent on their children, it's not possible to construct `ConfiguredNodes` for them directly as you traverse the graph, regardless of the direction you walk it. However there is still an algorithm that can construct the manifest with a single walk of the graph:

1. Walk the graph depth-first. While parsing children of graph nodes, pass down the graph node id so that we can construct edges (but not hook them up)
2. Construct and attach node-to-node local edges
3. Construct local edges for exposed ports, and attach them to the child nodes (there's no parent graph node yet). Pass these edges _up_ to the parent parsing function.
4. Connect any local edges passed up from child nodes.
5. Once all nodes are constructed (and local edges are connected), iterate over each node, resolving its local edges if its not a graph node.